requirements.txt
python-telegram-bot==21.0.1
shazamio==0.6.0
instaloader==4.11
moviepy==1.0.3
python-dotenv==1.0.1
ffmpeg-python==0.2.0
